requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
python-decouple==3.8
pyttsx3==2.90
requests==2.31.0
urllib3==2.0.7
